primitives: renamed TransactionDecoderError::Transitioning to Errored
Search-and-replace the error variant name, and update all the panic
conditions that are triggered by it. From a user point of view this
state does not represent "transitioning". It represents that we
called end() on a sub-decoder and that call yielded an error,
leaving us with no meaningful state.

The docs for Decoder::end and Decoder::current_chunk say that after
returning the initial error we are allowed to panic; update the panic
message to say that this is what we're doing. (We also have a panic
in read_limit, which we are not allowed to do, so replace that with
a dummy value return.)

No observable behavior changes except that some panic messages change.