Update mutant excludes
There are new mutants for cases that have previously been excluded due
to a difference in the path of the exclusion. There is also a timeout
where changing an = to != in push_bytes makes the decoding never end.

Make the exclusions more general to cover these cases.

Add an exclusion for the timeout.