Implement Decodable for Witness using encoding crate
Use the new `consensus_encoding` crate (by way of the `io` crate
decoding function) to implement `consensus_decode` for `Witness`.

This is ugly a bit ugly because we remove equality derives from
`ParseError` so that we can jam the witness decoding error into it.
Cleanly it should not go there but in order to keep
`encode::deserialize_partial` working we have to be able to return
a `ParseError` and we have to be able to include the witness decoding
error if there was one.

Each subsequent time we try to do this in other decoding impls we will
add a new error variant each time also. This encoding code is on its
way out and will hopefully be deprecated soon so I believe this is ok.