Add maintainer-tools support to the justfile
The justfile is a nice interface for local development, but ideally its
logic is the exact same as what runs in CI to minimize wasted cycles
found only in CI.

This patch introduces teh maintianer-tools-version file as the source
of truth for the version of CI script used in the workspace. The CI
tooling as well as the justfile have been updated to pull from it. The
justfile has also been extended with recipes which match the CI tasks
as close as possible.

Explicit rustup commands were introduced to the justfile with the new
"ci" recipe. This isn't expanding scope though since rustup was already
implicitly required by tasks like "format" which use the +toolchain
syntax.