Move encode_to_hash_engine to hashes from consensus_encoding
The encode_to_hash_engine function created a dependency on hashes from
consensus_encoding. Since consensus_encoding is moving to 1.0, it is
best to remove the dependency on the pre-1.0 hashes crate.

This patch moves the encode_to_hash_engine function to the hashes
crate, and reverses the dependency so that hashes now depends on
consensus_encoding instead of vice-versa.