Remove unused comment on Encodable trait implementation
Comments applied to function implementations of traits do not render
in compiled Rust documentation. The Encodable trait implementation
for BlockTransactionsRequest includes such a comment that also
duplicates the warning recorded on the indexes field of the struct.

This patch removes the useless comment above the function
consensus_encode in the Encodable implementation of
BlockTransactionsRequest.