Remove unused comment on BlockCheckedExt trait implementation
Comments applied to function implementations of traits do not render
in compiled Rust documentation. The BlockCheckedExt trait implementation
for Block<Checked> includes such a comment that also duplicates the
warning present on the trait definition.

This patch removes the useless comment above the function total_size
in the BlockCheckedExt implementation of Block<Checked>.