Remove unused comments on TransactionExt trait implementation
Comments applied to function implementations of traits do not render
in compiled Rust documentation. The TransactionExt trait implementation
for Transaction includes two such comments that duplicate similar
warnings present in the trait definition.

This patch removes the useless comments above the functions
total_size and base_size in the TransactionExt implementation of
Transaction.