Change extension trait function return types to Self
Within the extension traits in the bitcoin crate, some functions use
a Self return type, while others use the specific type name. Use of a
specific type name causes discontinuity in the codebase, is longer,
and is semantically less clear for distinguishing return of a trait
implementor's type vs an instance of a specific type.