Implement Write for [u8] Cursors
Currently, the Write trait is implemented for &mut [u8] byte slices.
The Cursor struct provides a convenient means of wrapping such slices
and accessing parts as needed. Implementing Write both for the Cursor
and the underlying slices provides a consistent interface for writing
whether using a Cursor or raw slice.

Implement the Write trait for Cursors on AsMut<[u8]> types.