TEMPORARY: Revert this patch after bitcoin-io v0.1.4 merges
We want CI to pass on the io-0.1.x branch but `bitcoin` is currently
borked and its breaking the build. Since we are only doing the `io`
release add a patch that fixes the build, after we build and tag the
patch before this one we can then revert this one. Ugly as sin I know.