consensus_encoding: implement batched allocation for vector decoders
Right now vector decoders (`WitnessDecoder`, `VecDecoder` and `ByteVecDecoder`) 
only checks that the element count/byte  < 4,000,000. This still allows an attacker
to claims a large size without providing the corresponding  data.

In this patch we allocates in 1 MB batches so an attacker now needs to provide
X MB of data to make us allocate X+1 MB of memory

The allocation is applied for WitnessDecoder, VecDecoder and ByteVecDecoder.