bitcoin: Change rust cfg lint to allow
We are trying to release `bitcoin-io` from this branch but currently
the build for `bitcoin` fails because of buggy feature gating on then
`units 0.1.x` branch.

```rust
        #[cfg(all(feature = "alloc", not(feature = "std")))]
        $crate::impl_tryfrom_str_from_int_infallible!(&str, $to, $inner, $fn; alloc::string::String, $to, $inner, $fn; alloc::boxed::Box<str>, $to, $inner, $fn);
        #[cfg(feature = "std")]
        $crate::impl_tryfrom_str_from_int_infallible!(&str, $to, $inner, $fn; std::string::String, $to, $inner, $fn; std::boxed::Box<str>, $to, $inner, $fn);
```

FTR this no longer exists on master. From above we see that in
`bitcoin` the macro will never be called for `--no-default-feature`
builds.

This doesn't matter on this branch so just change the lint to `allow`
instead of `deny`.

While we are at it add `fuzzing` to the list.