Clarify feature gate on impl_to_hex_from_lower_hex macro
The docs on internals::impl_to_hex_from_lower_hex macro state that the
calling crate must have an "alloc" feature. If this feature is not
enabled, the macro is a no-op. In practice, the macro has no such
feature gate, as it is primarily used in the bitcoin crate, which has
no "alloc" feature.

Adjust the wording of the internals::impl_to_hex_from_lower_hex macro
to clarify the true behaviour.