Create an empty crypto crate
In preparation for crate smashing the `crypto` module create an empty
crate so we can grab the name on crates.io.

I don't currently know whats the pros and cons of moving just keys to
moving everything in the `crypto` module. So the manifest description
is a bit vague.

FTR `bitcoin-keys` and `bitcoin_keys` are both taken on crates.io