Move ecdsa::SerializedSignature impls
In an effort to make the `ecdsa` and `taproot` modules look the same
so it is easier to tell exactly where they differ move some
`SelializedSignature` impls. Elect to use the layout in the `taproot`
one.

Except in `taproot` move the `Eq` impl to be grouped with the
`PartialEq` impls.