p2p: Lower initial allocation for Vec Decoder
In the implementation of consensus_decode_from_finite_reader for Vec<T>,
the initial vector capacity is calculated as 4_000_000 / 4 / sizeof<T>.
This is an arbitrary value that doesn't provide any specific benefit,
per discussion in #4797. Fedimint has a similar implementation of the
function using different, lower values.

Change the initial allocation capacity to match Fedimint's.