Only return valid data in sig slice
The `SerializedSignature` types have an internal buffer and we track
the length of valid data in the buffer. However in `AsRef` and friends
we are returning a slice of the whole buffer. This is confusing and a
potential footgun (eg `sig.as_ref()l.en()` isn't be meaningful).

Make the impls that get a byte slice reference only return valid data.
Done for both `ecdsa` and `taproot` type.