ecdsa: Add to/from_signature, len, and iter
To the `ecdsa` type add `to_signature`, `from_signature`, `len`, and
`iter` methods.

Change the implementation of `IntoIter` for a sig reference so we can
add `iter` method as done in `taproot`.

While we are at it fix up the docs and layout all the methods in the
same order for both types (`ecdsa` and `taproot`).

This unifies the APIs of the two types `ecdsa::SerializedSignature`
and `taproot::SerializedSignature`.