Allow underscores as separators in Amount values
Currently, the FromStr implementation on the Amount type has no means
of separating segments of the value, e.g. ','. Since different
countries utilise different standard separators, and '_' is common
internationally in various programming languages, it offers a suitable
option for separating segments of long number values.

Introduce _ separator handling in the Amount FromStr implementation,
disallowing the use of an '_' prefix, or consecutive '_' characters.
Also introduce a BadPosition error type for characters that are valid
in value input, but are in an invalid position.