primitives: Move MerkleNode and TxIdentifier into primitives
The MerkleNode trait in bitcoin provides convenience functions for
the TxMerkleNode and WitnessMerkleNode hash types to work with the
nodes as a tree. Moving the functionality of this trait into primitives
acts as a precursor to allow other functionality to move also.

Move the MerkleNode trait to primitives, making it private. Provide the
same interface for users through pub functions that wrap the private
trait. Introduce tests to cover mutants. Introduce TxIdentifier in
primitives.