primitives: Add consensus encoding traits for WitnessMerkleNode
The logic in compute_witness_commitment relies on the ability to
consensus encode TxMerkleNode and WitnessMerkleNode hashes into a
HashEngine type. Currently, only TxMerkleNode implements Encodable,
with the encoding for WitnessMerkleNode and TxMerkleNode implemented
in bitcoin::merkle_tree based on bitcoin::consensus.

Implement consensus_encoding::Encodable on WitnessMerkleNode to
match the current implementation on TxMerkleNode.