Move BlockUncheckedExt functions to Block<Unchecked> impl
During the move of Block into primitives, some functionality was left
in an extension trait BlockUncheckedExt, due to dependencies on other
logic in bitcoin. To simplify the API, this logic is better included
directly on the Block<Unchecked> type and the extension trait removed.

Remove the BlockUncheckedExt trait, and move validate,
compute_merkle_root, and compute_witness_commitment functions to the
Block<Unchecked> implementation. Move tests from bitcoin that only
require primitives to primitives.