Depend on primitives instead of bitcoin where type definitions overlap
Consolidate dependency tree for downstream crates that depend on
`bitcoin-p2p-messages` and `bitcoin-primitives` by reducing overlapping
dependencies.  For example, consider crate A which depends on
`bitcoin-primitives` and `bitcoin-p2p-messages`.  For types defined by
both `bitcoin-primitives` and `bitcoin`, Crate A may still need to
depend on `bitcoin` if a type of `bitcoin-p2p-messages` depends on
`bitcoin` instead of `bitcoin-primitives`.