Introduce wrapper type for secp256k1::Keypair
In order to solve the issue of imprecise types for from_str in
XOnlyPublicKey, a wrapper type was implemented, alongside a
corresponding narrow error type. This solution provided a cleaner
interface for such keys, in place of directly working with secp256k1
XOnlyPublicKey types.

Add a wrapper type for secp256k1::Keypair to tighten the error type
for from_str.