Use usize for CompactSizeDecoder and add new_with_limit
Following the change to CompactSizeEncoder::new to take a usize, the
return types and ranges of the CompactSizeEncoder and CompactSizeDecoder
now differ. Since a device's memory addressing is limited to usize,
returning values beyond that size is largely useless. Converting the
CompactSizeDecoder to return a usize also allows the
cast_to_usize_if_valid function to be merged into the decoder itself.

Convert the CompactSizeDecoder to return a usize. Remove
cast_to_usize_if_valid and introduce its functionality into
CompactSizeDecoder.