Add pop to ArrayVec
In order for ArrayVec to match the API of alloc::vec::Vec, it requires
both pop() and push() in order to correctly emulate a stack. Currently,
ArrayVec only implements push(), and so is unable to replace Vec in
features that require stack functionality.

Add pop() function to ArrayVec to return last item as an Option.