units: add `too_many_lines` clippy whitelist to `parse_signed_to_satoshi`
After the format bot runs, this function (which is currently just below
the max line threshold) tips over, meaning that our format PRs are not
passing CI.

Possibly this function should be split up, but I don't think so. It's
split up into a few logical sections which are isolated from each other
but which wouldn't benefit from the even-further-separation that would
happen if we added extra functions.

Anyway we can debate that in a separate issue if people want. For now we
need the format job to pass CI, since it's a 700+ line diff already and
growing every week.