hashes: silence new `unused-assignments` clippy lint
In our `round!` macro invocation for the implementation for sha256 and
sha512 we do an assignment of a variable `$w` which is fed into the next
round. On the last round, this assignment is unused, which makes clippy
unhappy on recent nightlies.

Fix this so we can update our nightly compiler.