internals: Bump version to 0.4.2
In preparation for releasing `bitcoin 0.33.0-rc.0` we need do a point
release of `internals` because we move a macro out of `bitcoin` and
over to `internals`. No other crate uses the macro so far so we only
update the `bitcoin` dependency on `internals` leaving the other
crates as they are.