EOL bitcoin v0.30.x
We are currently only committed to supporting the latest 3 versions of
`bitcoin`. As of today that is `v0.31.x`, `v0.32.x`, and `v0.33.x`.

`0.33.x` is currently released as `v0.33.0-beta`.

End of Life this major version by doing:

- Document in CHANGELOG as well as the crate root module.
- Bump version number.
- Manually edit the minimal/recent lock files.

Tested minimal lock file by copying it into place and running `cargo
--locked check`. No further testing done since there are no code
changes in this patch anyways.

The changelog is already present and up to date.