Move `FilterHash`, `FilterHeader` to `p2p`
These types are defined in
[BIP-157](https://github.com/bitcoin/bips/blob/master/bip-0157.mediawiki#specification).

Defining these types within `p2p` is appropriate, as these are used for
wire messages. By doing this move 1. no other crates would depend on
`bip158` 2. we are adhering to the policy that a crate with a release
that matches the BIP must implement only that BIP. As part of this move,
a single test `assert` is removed, but this was testing a method that
didn't make much sense in the first place. If a user wants to get a
filter header, they should use the filter hash and filter header
directly.