internals: Add a changelog for the 0.4.1 release
In Nashville, in the work to get `primitives` out we rushed an
`internals` minor release that was actually breaking and should have
been a major release. For some reason at that time I omitted a
changelog, in hindsite now I look at the diff to `macros` I am not
surprised I missed the breaking change.

  git diff internals-0.4.0 bitcoin-internals-0.4.1 -- internals/src/macros.rs

Add a changelog for the `internals v0.4.1` release. Note in it that
the release violated semver rules.