psbt: Fix tapscript witness signature ordering
BIP-0342 requires that when executing a spend of a P2TR output using the
script path, the witness signatures are ordered in reverse of the order
in which their corresponding public keys appear in the script. For
example, a script with `pubkey_1` CHECKSIG `pubkey_2` CHECKSIGADD ...
`pubkey_n` CHECKSIGADD must have a witness ordered as `w_n`, ..., `w_2`,
`w_1`, where `w_i` corresponds to `pubkey_i`. This is necessary because
of the LIFO nature of stack execution. The last element pushed is the
first popped, so the signature for `pubkey_1` must be pushed last.

Previously, witness signatures were pushed onto the witness stack in
sorted key order, which may not have matched the order required by the
specification, leading to `Invalid Schnorr signature` errors.  This
commit fixes the `finalize_taproot_script_path_input` function to  match
the specification.

Reference:
- BIP-0342: https://github.com/bitcoin/bips/blob/master/bip-0342.mediawiki