io: Bump version to 0.4.0-rc.0
We updated a bunch of dependencies and forgot to re-release
`bitcoin-io`. And this is breaking the attempt at releasing `bitcoin
0.33.0-beta.0`.

Use RC because we have a dependency on `consensus_encoding` RC release
that uses the `=` syntax to pin the exact version. Therefore everytime
we do a new RC release of that crate we have to do `io` as well.

In preparation for release bump the version, add a changelog entry,
and update the lockfiles. Also update the version everywhere `io` is
used as a dependency including adding an explicit version number to
`p2p` (which will be needed to release it).