Add PRNG to `MerkleBlock` tests
This is a PRNG that can replace the dependency on `rand` in
this unit test. Even if this test failed in CI, I'm not sure how we
would actually reproduce it. The current state of the test seems more
apt for a fuzzing target where there is tooling to log and reproduce
failing cases. This would allow for determinism in the test and could be
used as a sanity check when making changes to `MerkleBlock`

Co-authored-by: apoelstra <apoelstra@wpsoftware.net>